creating pages module to define and manage browser windows deleted unused run-edchain.sh updated the close button for child windows moved icon.icns to build directory
