+++ package.json
@@ -8,10 +8,10 @@
   "main": "src/app.js",
   "scripts": {
     "start": "electron src/app.js",
-    "build": "electron-packager . edChain --out dist/ --overwrite --ignore=electron-packager --icon=./src/public/img/icon.icns",
+    "build": "electron-packager . edChain --out dist/ --overwrite --ignore=electron-packager --icon=./build/icon.icns",
     "rpm64": "electron-installer-redhat --src dist/edChain-linux-x64/ --dest dist/installers/ --arch x86_64 --overwrite",
     "deb64": "electron-installer-debian --src dist/edChain-linux-x64/ --dest dist/installers/ --arch amd64 --overwrite",
-    "darwin": "electron-installer-dmg dist/edChain-darwin-x64/ edchain --out dist/installers/ --overwrite --icon=./src/public/img/icon.icns"
+    "darwin": "electron-installer-dmg dist/edChain-darwin-x64/ edchain --out dist/installers/ --overwrite --icon=./build/icon.icns"
   },
   "devDependencies": {
     "electron": "^1.7.9",
