+++ src/app.js
@@ -1,20 +1,22 @@
 var path = require('path');         // https://nodejs.org/api/path.html
 var url = require('url');           // https://nodejs.org/api/url.html
-var log = require('electron-log');
-const { exec } = require('child_process');
+
 const platform = require('os').platform();
+
 const { 
-    ipcMain, app, protocol, BrowserWindow, 
-    Menu, Tray, nativeImage 
+    ipcMain, app, protocol 
 } = require('electron');
 
-var __windows = {};
 
-var __logSubscribers = {};
+const pages = require('./pages.js');
 
+// this is actually to ping the different windows
+// sucks but thats how you do it.
+// this is probably how all windows should be connected to 
+// the main process.
 var ipfs = require('./process_ipfs')({
     "afterLogUpdateHook": function(ipfsLog){
-        for(let subscriber in __logSubscribers){
+        for(let subscriber in pages.getLogSubscribers()){
             if(typeof subscriber.send === 'function'){
                 subscriber.send('ipfs:logging', ipfsLog);
             }
@@ -36,166 +38,10 @@ var registerListeners = function(listeners){
 
 registerListeners({ipfs});
 
-// TODO: MOVE
-var getIcon = (function(){
-    var iconMap = {
-        "darwin": "/public/img/icon.icns",
-        "linux": "/public/img/icon.png"
-    };
-
-    return function(){
-        return iconMap[platform] || "/public/img/icon.png";
-    };
-})();
-
-// TODO: MOVE
-var getTray = (function(){
-    var 
-        tray = null, 
-        imageFolder = path.resolve(__dirname, "public/img/");
-
-    return function(__window){
-
-        if(!tray){
-            // http://electron.rocks/proper-tray-icon/
-            if (platform === 'darwin' || platform === 'linux') {  
-                trayImage = path.resolve(imageFolder, 'icon.png');
-            }
-            else if (platform === 'win32') {  
-                trayImage = path.resolve(imageFolder, 'icon.ico');
-            }
-            tray = new Tray(path.resolve(__dirname, trayImage));
-            if (platform === "darwin") {  
-                tray.setPressedImage(path.resolve(imageFolder, 'icon.png'));
-            }
-        }
-
-        tray.on('click', () => {
-            __window.isVisible() ? __window.hide() : __window.show();
-        });
-
-        __window.on('show', () => {
-            tray.setHighlightMode('always');
-        });
-
-        __window.on('hide', () => {
-            tray.setHighlightMode('never');
-        });
-        return tray;
-    };
-})();
-// TODO: MOVE
-var createChildWindow = function (mainWindow, url) {
-    
-    var child = createWindow({
-        parent: mainWindow, 
-        modal:true, 
-        show:true,
-        hasIpfsLogging: true
-    });
-
-    if (process.platform === 'darwin') {
-        child.webContents.once("did-navigate", function(event, ...args){
-            child.webContents.once("dom-ready", function(event, ...args){
-                // TODO: THIS IS GOING TO NEED TO BE MORE EXTENSIVE
-                // need to wrap more than just close. also need to inercept events
-                var pageModification = `(function(){
-                    var $nav = $('<nav><button id="close-sheet">Close</button></nav>').prependTo('body');
-                    $nav.on("click", "#close-sheet", function(event){
-                        event.preventDefault();
-                        window.close();
-                    });
-                })();`;
-                event.sender.executeJavaScript(pageModification, null, function(){log.info("here")});
-            });
-        });
-    }
-    child.loadURL(url);
-    return child;
-
-};
-
-
-// TODO: MOVE
-var showChildWindow = function(browserWindow){
-    browserWindow.show();
-    // browserWindow.openDevTools();
-    // log.info(browserWindow.webContents); 
-};
-
-// TODO: MOVE
-// Window Factory
-var createWindow = function createWindow(config){
-    var 
-        __id, browserWindow, 
-        hasIpfsLogging = false;
-    
-    if(Object.hasOwnProperty.call(config, "hasIpfsLogging")){
-        hasIpfsLogging = config["hasIpfsLogging"];
-        config["hasIpfsLogging"] = null;
-    }
-
-    browserWindow = new BrowserWindow(config);
-    __id = browserWindow.id;
-    
-    __windows[__id] = browserWindow;
-    
-    if(hasIpfsLogging){
-        __logSubscribers[__id] = browserWindow.webContents;
-    }
-
-    browserWindow.on('closed', function(){
-        __windows[__id].removeAllListeners();
-        
-        if(__logSubscribers[__id]) __logSubscribers[__id] = null;
-
-        __windows[__id] = null;
-    });
-    return browserWindow;
-};
-
-// TODO: MOVE
-var createMainWindow = function createMainWindow(){
-    var
-        settingsWindow, mainWindow;
-
-    mainWindow = createWindow({
-        width: 960,
-        height: 540,
-        //frame: false,
-        icon: path.resolve(__dirname, "public/img/icon.png")
-    });
-
-    mainWindow.tray = getTray(mainWindow);
-
-    mainWindow.loadURL(`file://${__dirname}/public/html/index.html`);
-    
-    mainWindow.once('ready-to-show', () => {
-        mainWindow.show();
-    });
-   
-    ipcMain.on('createAndShowChildWindow', function(event,url){
-        showChildWindow(createChildWindow(mainWindow, url));
-    });
-
-    ipcMain.on('createChildWindow', function(event, url){
-        settingsWindow = createChildWindow(mainWindow,url);
-    });
-
-    ipcMain.on('showChildWindow', function(){
-        showChildWindow(settingsWindow);
-    });
-
-    ipcMain.on("closePage", function(event, id){
-        __windows[id].close();
-    });
-};
-
 // macOS
 // https://electronjs.org/docs/api/app#appdockseticonimage-macos
 if (platform === "darwin"){
     // Seems to hate my .icns
-    // app.dock.setIcon(__dirname + getIcon());
     app.dock.setIcon(path.resolve(__dirname, "public/img/icon.png"));
 }
 
@@ -212,7 +58,7 @@ app.on('ready', function(){
     //     if (err) console.error('Failed to register protocol')
     // });
 
-    createMainWindow();
+    pages.createMainWindow();
 
 });
 
@@ -233,8 +79,9 @@ app.on('quit', () => {
 app.on('activate', (event, hasVisibleWindows) => {
     // On macOS it's common to re-create a window in the app when the
     // dock icon is clicked and there are no other windows open.
-    if (!Object.keys(__windows).length) {
-        createMainWindow();
+    // move
+    if (!pages.hasWindows()) {
+        pages.createMainWindow();
     } else if (hasVisibleWindows) {
         event.preventDefault();
     }
