+++ src/public/html/settings.html
@@ -32,9 +32,6 @@
     </head>
     <body>
         <div class="container">
-            <button type="button" class="close" id="close-window" aria-label="Close">
-                <span aria-hidden="true">&times;</span>
-            </button>
             <nav class="nav nav-tabs" role="tablist">
                 <a class="nav-item nav-link" data-toggle="tab" id="logs-tab" href="#logs-pane" role="tab">logs</a>
                 <a class="nav-item nav-link active" data-toggle="tab" id="settings-tab" href="#settings-pane">settings</a>
